backend/requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
django-anymail==7.0.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-material==1.6.0
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.8.1
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.8
jsmin==2.2.2
ldif3==3.1.1
Markdown==3.1.1
Pillow==6.2.1
pprint==0.1
promise==2.3
psycopg2==2.8.4
python-card-me==0.9.3
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
Rx==1.6.1
sib-api-v3-sdk==6.0.3
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.25.7
